adds olives (for olive oil), fire stuff

+++ code/modules/reagents/reagent_containers/glass.dm
@@ -79,6 +79,10 @@
 		if (reagents.total_volume && !istype(src, /obj/item/weapon/reagent_containers/glass/small_pot))
 			playsound(src,'sound/effects/Splash_Small_01_mono.ogg',50,1)
 			user << "<span class='notice'>You splash the solution onto [target].</span>"
+			if (reagents.has_reagent("petroleum", 5))
+				new/obj/effect/decal/cleanable/blood/oil(src.loc)
+			else if (reagents.has_reagent("olive_oil", 10))
+				new/obj/effect/decal/cleanable/blood/oil(src.loc)
 			reagents.splash(target, reagents.total_volume)
 			return
 
@@ -102,6 +106,19 @@
 
 			user << "You smash the grapes, producing grapejuice."
 			reagents.add_reagent("grapejuice", 5)
+			qdel(W)
+			return
+		if (istype(W, /obj/item/weapon/reagent_containers/food/snacks/grown/olives))
+
+			if (!is_open_container())
+				user << "<span class='notice'>\The [src] is closed.</span>"
+				return
+			if (!reagents.get_free_space())
+				user << "<span class='notice'>[src] is full.</span>"
+				return
+
+			user << "You smash the olives, producing olive oil."
+			reagents.add_reagent("olive_oil", 3)
 			qdel(W)
 			return
 	proc/update_name_label()
